Renamed toolstep files for clarity
